Cleaned up removing unused
