+++ src/servers/qos.rs
@@ -1,9 +1,10 @@
-use crate::{constants::QOS_PORT, net::public_address, show_error};
+use crate::{constants::QOS_PORT, show_error};
 use std::{
     net::{IpAddr, Ipv4Addr, SocketAddr},
     process::exit,
+    time::{Duration, SystemTime},
 };
-use tokio::net::UdpSocket;
+use tokio::{net::UdpSocket, sync::RwLock};
 
 /// Starts the Quality Of Service server which handles providing the public
 /// address values to the clients that connect.
@@ -70,4 +71,82 @@ async fn get_address(addr: &SocketAddr) -> Option<Ipv4Addr> {
         return Some(value);
     }
     None
+}
+
+/// Caching structure for the public address value
+enum PublicAddrCache {
+    /// The value hasn't yet been computed
+    Unset,
+    /// The value has been computed
+    Set {
+        /// The public address value
+        value: Ipv4Addr,
+        /// The system time the cache expires at
+        expires: SystemTime,
+    },
+}
+
+/// Cache value for storing the public address
+static PUBLIC_ADDR_CACHE: RwLock<PublicAddrCache> = RwLock::const_new(PublicAddrCache::Unset);
+
+/// Cache public address for 30 minutes
+const ADDR_CACHE_TIME: Duration = Duration::from_secs(60 * 30);
+
+/// Retrieves the public address of the server either using the cached
+/// value if its not expired or fetching the new value from the one of
+/// two possible APIs
+async fn public_address() -> Option<Ipv4Addr> {
+    {
+        let cached = &*PUBLIC_ADDR_CACHE.read().await;
+        if let PublicAddrCache::Set { value, expires } = cached {
+            let time = SystemTime::now();
+            if time.lt(expires) {
+                return Some(*value);
+            }
+        }
+    }
+
+    // Hold the write lock to prevent others from attempting to update aswell
+    let cached = &mut *PUBLIC_ADDR_CACHE.write().await;
+
+    // API addresses for IP lookup
+    let addresses = ["https://api.ipify.org/", "https://ipv4.icanhazip.com/"];
+    let mut value: Option<Ipv4Addr> = None;
+
+    // Try all addresses using the first valid value
+    for address in addresses {
+        let response = match reqwest::get(address).await {
+            Ok(value) => value,
+            Err(_) => continue,
+        };
+
+        let ip = match response.text().await {
+            Ok(value) => value.trim().replace('\n', ""),
+            Err(_) => continue,
+        };
+
+        if let Ok(parsed) = ip.parse() {
+            value = Some(parsed);
+            break;
+        }
+    }
+
+    // If we couldn't connect to any IP services its likely
+    // we don't have internet lets try using our local address
+    if value.is_none() {
+        if let Ok(IpAddr::V4(addr)) = local_ip_address::local_ip() {
+            value = Some(addr)
+        }
+    }
+
+    let value = value?;
+
+    // Update cached value with the new address
+
+    *cached = PublicAddrCache::Set {
+        value,
+        expires: SystemTime::now() + ADDR_CACHE_TIME,
+    };
+
+    Some(value)
 }
